Implement authorization usung Gate method

+++ app/Http/Controllers/QuestionController.php
@@ -67,6 +67,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Acess denied!!");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -79,6 +82,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Acess denied!!");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been upadated');
@@ -92,6 +99,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Acess denied!!");
+        }
+
         $question->delete();
 
         return redirect('/questions')->with('success', "Your message has been deleted successfuly!!");
